Changed no more needed name mangling
